+++ src/measures/BooleanMeasure.hpp
@@ -8,12 +8,12 @@ public:
     virtual ~BooleanMeasure();
     virtual double eval(const Alignment& A) = 0;
 
-    virtual int getIncChangeOp(uint peg, uint oldHole, uint newHole, const Alignment& A);
-    virtual int getIncSwapOp(uint peg1, uint peg2, uint hole1, uint hole2, const Alignment& A);
+    virtual double getIncChangeOp(uint peg, uint oldHole, uint newHole, const Alignment& A);
+    virtual double getIncSwapOp(uint peg1, uint peg2, uint hole1, uint hole2, const Alignment& A);
 
 protected:
-    int computeIncChangeOp(uint peg, uint oldHole, uint newHole, const Alignment& A);
-    int computeIncSwapOp(uint peg1, uint peg2, uint hole1, uint hole2, const Alignment& A);
+    double computeIncChangeOp(uint peg, uint oldHole, uint newHole, const Alignment& A);
+    double computeIncSwapOp(uint peg1, uint peg2, uint hole1, uint hole2, const Alignment& A);
 };
 
 #endif
